refactor(rooms): HS-158-05a the extraction — ProjectMemoryCore moves into the project-room boundary, nothing changes
The 751-line core becomes web/src/features/project-room/: api.ts
(nine typed wrappers around the existing fetches), model.ts (wire
types, composeProjectTimeline, decode helpers — WEB-ARC-004),
useProjectRoomController.ts (discriminated loadStatus, WEB-ARC-003),
ProjectRoomCore.tsx (identical JSX, same testids, same CSS).
ProjectMemoryCore.tsx is now a 16-line compatibility re-export; the
lazy loader and every existing consumer resolve through it. Zero
behavior change proven: the 157 pins and the whole existing suite
pass UNCHANGED (zero test-file edits), typecheck clean, architecture
guard clean; +21 new unit tests for model/controller. Story 05 stays
in-progress — the adoption half waits for GET /room (04).

PMO-Story: HS-158-05
PMO-Contract-Digest: sha256:649f4656ffb3db194243932ddc0eacbb6cfa508773521cbbc8680cbe6d49df83
